requirements.txt
Django~=4.1.1
celery~=5.2.7
djangorestframework~=3.14.0
requests~=2.28.1
SQLAlchemy~=1.4.41
eventlet~=0.33.1
amqp~=5.1.1
